refactor(storage/querybuilder): move query building logic to own package
